+++ app/schemas.py
@@ -89,10 +89,10 @@ class PokerTableMeta(BaseModel):
     club_id: int
     created_by_user_id: int
     max_seats: int
-    small_blind: int
-    big_blind: int
+    small_blind: float
+    big_blind: float
     bomb_pot_every_n_hands: Optional[int]
-    bomb_pot_amount: Optional[int]
+    bomb_pot_amount: Optional[float]
     status: str
     created_at: datetime
 
@@ -106,10 +106,10 @@ class ClubDetail(ClubRead):
 
 class ClubTableCreate(BaseModel):
     max_seats: int = 6
-    small_blind: int = 1
-    big_blind: int = 2
+    small_blind: float = 1
+    big_blind: float = 2
     bomb_pot_every_n_hands: Optional[int] = None
-    bomb_pot_amount: Optional[int] = None
+    bomb_pot_amount: Optional[float] = None
 
 
 # ---------- Poker Engine ----------
@@ -118,8 +118,8 @@ class PlayerState(BaseModel):
     id: int
     name: str
     seat: int
-    stack: int
-    committed: int
+    stack: float
+    committed: float
     in_hand: bool
     has_folded: bool
     all_in: bool
@@ -131,16 +131,16 @@ class TableState(BaseModel):
     id: int
     hand_number: int
     street: str
-    pot: int
+    pot: float
     board: List[str]
-    current_bet: int
+    current_bet: float
     next_to_act_seat: Optional[int]
     action_deadline: Optional[float]
     dealer_button_seat: Optional[int]
     small_blind_seat: Optional[int]
     big_blind_seat: Optional[int]
-    small_blind: int
-    big_blind: int
+    small_blind: float
+    big_blind: float
     players: List[PlayerState]
 
     model_config = ConfigDict(from_attributes=True)
@@ -149,10 +149,10 @@ class TableState(BaseModel):
 class CreateTableRequest(BaseModel):
     club_id: int
     max_seats: int = 6
-    small_blind: int = 1
-    big_blind: int = 2
+    small_blind: float = 1
+    big_blind: float = 2
     bomb_pot_every_n_hands: Optional[int] = None
-    bomb_pot_amount: Optional[int] = None
+    bomb_pot_amount: Optional[float] = None
 
 
 class CreateTableResponse(BaseModel):
